chore: drop and clean from phase 1
