+++ codex-rs/core/src/memories/mod.rs
@@ -4,16 +4,20 @@
 //! - Phase 1: select rollouts, extract stage-1 raw memories, persist stage-1 outputs, and enqueue consolidation.
 //! - Phase 2: claim a global consolidation lock, materialize consolidation inputs, and dispatch one consolidation agent.
 
+mod dispatch;
+mod phase1;
+mod phase2;
 pub(crate) mod prompts;
-mod stage_one;
-mod startup;
+mod start;
 mod storage;
-
 #[cfg(test)]
 mod tests;
 
-use std::path::Path;
-use std::path::PathBuf;
+/// Starts the memory startup pipeline for eligible root sessions.
+/// This is the single entrypoint that `codex` uses to trigger memory startup.
+///
+/// This is the entry point to read and understand this module.
+pub(crate) use start::start_memories_startup_task;
 
 mod artifacts {
     pub(super) const ROLLOUT_SUMMARIES_SUBDIR: &str = "rollout_summaries";
@@ -22,10 +26,12 @@ mod artifacts {
 
 /// Phase 1 (startup extraction).
 mod phase_one {
+    /// Prompt used for phase 1.
+    pub(super) const PROMPT: &str = include_str!("../../templates/memories/stage_one_system.md");
     /// Maximum number of rollout candidates processed per startup pass.
     pub(super) const MAX_ROLLOUTS_PER_STARTUP: usize = 64;
     /// Concurrency cap for startup memory extraction and consolidation scheduling.
-    pub(super) const CONCURRENCY_LIMIT: usize = MAX_ROLLOUTS_PER_STARTUP;
+    pub(super) const CONCURRENCY_LIMIT: usize = 64;
     /// Fallback stage-1 rollout truncation limit (tokens) when model metadata
     /// does not include a valid context window.
     pub(super) const DEFAULT_STAGE_ONE_ROLLOUT_TOKEN_LIMIT: usize = 150_000;
@@ -46,6 +52,8 @@ mod phase_one {
     pub(super) const JOB_LEASE_SECONDS: i64 = 3_600;
     /// Backoff delay (seconds) before retrying a failed stage-1 extraction job.
     pub(super) const JOB_RETRY_DELAY_SECONDS: i64 = 3_600;
+    /// Maximum number of threads to scan.
+    pub(super) const THREAD_SCAN_LIMIT: usize = 5_000;
 }
 
 /// Phase 2 (aka `Consolidation`).
@@ -74,6 +82,9 @@ mod metrics {
     pub(super) const MEMORY_PHASE_TWO_INPUT: &str = "codex.memory.phase2.input";
 }
 
+use std::path::Path;
+use std::path::PathBuf;
+
 pub fn memory_root(codex_home: &Path) -> PathBuf {
     codex_home.join("memories")
 }
@@ -89,8 +100,3 @@ fn raw_memories_file(root: &Path) -> PathBuf {
 async fn ensure_layout(root: &Path) -> std::io::Result<()> {
     tokio::fs::create_dir_all(rollout_summaries_dir(root)).await
 }
-
-/// Starts the memory startup pipeline for eligible root sessions.
-///
-/// This is the single entrypoint that `codex` uses to trigger memory startup.
-pub(crate) use startup::start_memories_startup_task;
